backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pymongo>=4.5.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
pydantic>=1.10.0
ollama>=0.1.0